pyproject.toml
[build-system]
requires = ["flit"]
build-backend = "flit.buildapi"

[project]
name =  "aiida-amneris"
authors = [
    {name = "Jusong Yu", email = "jusong.yu@psi.ch"},
]
description = "Sample Python Project for creating a new Python Module"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11"
]
requires-python = ">=3.8.1"
dynamic = ["version"]
dependencies = [
    "aiida-core~=2.5",
    "solara~=1.25",
    "humanfriendly~=10.0",
    "ipywidgets~=8.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit~=3.6",
    "pytest~=7.4",
]

[project.urls]
Documentation = "https://github.com/unkcpz/aiida-amneris#readme"
Source = "https://github.com/unkcpz/aiida-amneris"
Tracker = "https://github.com/unkcpz/aiida-amneris/issues"

[tool.flit.module]
name = "aiida_amneris"
